rboot: port upstream rcore-os/rboot 2026 changes onto Eclipse's tree
Our rboot is upstream's 2022 tree plus Eclipse's additions (EDID, splash
and progress bar, multi-GOP selection, raw ExitBootServices with retries,
diagnostic IDT, chunked file loading, tolerant mappings). This brings in
what upstream did since, keeping all of that:

* uefi 0.16/0.26 -> 0.40 API (free `boot::`/`system::` functions, no
  SystemTable/BootServices handles, `#[entry] fn efi_main() -> Status`,
  `helpers::init`), xmas-elf 0.10, edition 2024, version 0.2.0. The
  x86_64 crate comes from crates.io with only `instructions` +
  `abi_x86_interrupt` (no `step_trait`), and rboot moves to the workspace
  nightly (2026-09-01) without build-std: the rustup-provided UEFI
  targets are used directly.
* `arch/` split: `page_table.rs` becomes `arch/x86_64.rs` (plus
  current_page_table, the UEFI frame allocator and the hand-off), and
  upstream's `arch/aarch64.rs` is imported verbatim: ELF loading to
  physical memory, boot page tables from the UEFI memory map and the
  EL2->EL1 / MMU hand-off. `Aarch64BootInfo` and the `uart_base`,
  `gic_base`, `firmware_type` config keys come with it; `make build
  TARGET=aarch64-unknown-uefi` produces an aarch64 rboot.efi.
* Config: `0x`/`0X` numeric parsing, lines without `=` ignored, missing
  rboot.conf falls back to the built-in defaults; a missing kernel at
  `kernel_path` tries `\os`, `\EFI\Boot\os`, `\EFI\zCore\zcore.elf`; a
  missing initramfs and absent ACPI/SMBIOS tables no longer abort the
  boot; a failed GOP `set_mode` keeps the current mode.
* `BootInfo::memory_map` is `Vec<MemoryDescriptor>` by value (upstream's
  layout), copied into pre-allocated capacity after ExitBootServices so
  nothing allocates once boot services are gone; the console logger is
  switched off right before that call.
* `wcslen` shim: newer LLVM turns uefi-rs' `CStr16` length loops into a
  libc `wcslen` call, which no UEFI target provides; a volatile-loop
  definition satisfies the link.
* kernel-hal follows to uefi 0.40 (`uefi::boot::{MemoryType,
  MemoryDescriptor}`, memory map slice by value) so the kernel and the
  bootloader agree on the `BootInfo` types.
* example-kernel refreshed from upstream (aarch64 test flow), README
  rewritten with the merged option list and the Eclipse additions.

Verified: `make build` and `make clippy` for both UEFI targets, `cargo
fmt --check`.

+++ rboot/src/idt.rs
@@ -21,9 +21,9 @@ static mut IDT: InterruptDescriptorTable = InterruptDescriptorTable::new();
 pub fn init(mode: ModeInfo, fb_addr: u64) {
     let (sw, sh) = mode.resolution();
     FB_ADDR.store(fb_addr, Ordering::SeqCst);
-    STRIDE.store(mode.stride() as usize, Ordering::SeqCst);
-    SW.store(sw as usize, Ordering::SeqCst);
-    SH.store(sh as usize, Ordering::SeqCst);
+    STRIDE.store(mode.stride(), Ordering::SeqCst);
+    SW.store(sw, Ordering::SeqCst);
+    SH.store(sh, Ordering::SeqCst);
 
     unsafe {
         let idt = &mut *core::ptr::addr_of_mut!(IDT);
